+++ src/store.js
@@ -1,6 +1,22 @@
-import { createStore } from "vuex";
+import { createStore } from 'vuex'
 
-const store = createStore({});
+const store = createStore({
+    state: {
+        data: null
+    },
+    mutations: {
+        setData(state, data) {
+            state.data = data;
+        }
+    },
+    actions: {
+        setData({ commit }, data) {
+            commit('setData', data);
+        }
+    },
+    getters: {
+        getData: state => state.data
+    }
+})
 
-
-export default store;
+export default store
